+++ packages/annotation-toolkit/src/index.js
@@ -3,12 +3,20 @@ import AppShell from "./AppShell";
 import Layer, { LayerContext } from "./layers/Layer";
 import BBoxFrame from "./layers/BBoxFrame";
 import VideoPlayer from "./layers/VideoPlayer";
-import MovableRect from "./layers/MovableRect";
+import MovableRect, { getInterpolatedFrames } from "./layers/MovableRect";
 
 import "./react-mosaic-component.css";
 import "./index.css";
 import "./layers/RRRR/react-rect.css";
 
-export { AppShell, Layer, LayerContext, BBoxFrame, VideoPlayer, MovableRect };
+export {
+  AppShell,
+  Layer,
+  LayerContext,
+  BBoxFrame,
+  VideoPlayer,
+  MovableRect,
+  getInterpolatedFrames,
+};
 
 export * from "./helpers";