+++ packages/annotation-toolkit/release-notes.md
@@ -8,9 +8,10 @@
 - **ENHANCEMENT** - VideoPlayer playing can be toggled by clicking on the video.
 - **ENHANCEMENT** - Add a getLabel property to MovableRect that adds a label to the box.
 
-- **TODO** - Add finalizeData helper to allow exporting interpolated frame data
+- **NEW** - Add `getInterpolatedFrames(allKeyframes)` helper to convert keyframes into interpolated frames, frame-by-frame. Returns array of: `{frame, value: [x, y, width, height]}`
+
 - **TODO** - Add greater css specificity for css styles to avoid conflicts
-- **TODO** - refactor reused code
+- **TODO** - Refactor reused code
 - **TODO** - Confirm that no more props need to be extracted
 ---
 - **BREAKING** - Layer actions now require a render prop instead of React node.